separate preview logic from tests
